Using the context Api to manage state
